ajout des tokens dans les controllers category card collection

+++ src/main/java/com/projetb3/api/controller/CategoryController.java
@@ -9,6 +9,8 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Optional;
 
+import static com.projetb3.api.security.AuthenticationWithJWT.verifier;
+
 @Controller
 @RequestMapping("/categories")
 public class CategoryController {
@@ -34,40 +36,48 @@ public class CategoryController {
         return ResponseEntity.notFound().build();
     }
 
-    @PostMapping
-    public ResponseEntity<String> create(@RequestBody Category category) {
-        categoryService.save(category);
-        return ResponseEntity.ok().body("La catégorie a été créée.");
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable("id") final int id) {
-        Optional<Category> optCategory = categoryService.get(id);
-        if (optCategory.isPresent()) {
-            categoryService.delete(id);
-            return ResponseEntity.ok().body("La catégorie a été supprimée.");
-        }
-        return ResponseEntity.notFound().build();
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<String> update(@PathVariable("id") final int id, @RequestBody Category modified) {
-        Optional<Category> optCategory = categoryService.get(id);
-        if (optCategory.isPresent()) {
-            Category current = optCategory.get();
-            if (modified.getName() != null) {
-                current.setName(modified.getName());
-            }
-            categoryService.save(current);
-            return ResponseEntity.ok().body("La catégorie " + current.getId() + " a été modifiée.");
-        }
-        return ResponseEntity.notFound().build();
-    }
-
     @GetMapping("/piece={id_room}")
     public ResponseEntity<List<Category>> getCategoriesOfRoom(@PathVariable("id_room") final int id_room) {
         List<Category> listCategories = categoryService.getCategoriesOfRoom(id_room);
         return ResponseEntity.ok(listCategories);
     }
 
+    @PostMapping
+    public ResponseEntity<String> create(@RequestBody Category category, @RequestHeader("Authentication") final String token) {
+        if (verifier(token, Optional.empty())) {
+            categoryService.save(category);
+            return ResponseEntity.ok().body("La catégorie a été créée.");
+        }
+        return ResponseEntity.badRequest().build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> delete(@PathVariable("id") final int id, @RequestHeader("Authentication") final String token) {
+        if (verifier(token, Optional.empty())) {
+            Optional<Category> optCategory = categoryService.get(id);
+            if (optCategory.isPresent()) {
+                categoryService.delete(id);
+                return ResponseEntity.ok().body("La catégorie a été supprimée.");
+            }
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.badRequest().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<String> update(@PathVariable("id") final int id, @RequestBody Category modified, @RequestHeader("Authentication") final String token) {
+        if (verifier(token, Optional.empty())) {
+            Optional<Category> optCategory = categoryService.get(id);
+            if (optCategory.isPresent()) {
+                Category current = optCategory.get();
+                if (modified.getName() != null) {
+                    current.setName(modified.getName());
+                }
+                categoryService.save(current);
+                return ResponseEntity.ok().body("La catégorie " + current.getId() + " a été modifiée.");
+            }
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.badRequest().build();
+    }
 }
