+++ src/main/java/com/projetb3/api/controller/CollectionController.java
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
 
+import static com.projetb3.api.security.AuthenticationWithJWT.verifier;
+
 @Controller
 @RequestMapping("/collections")
 public class CollectionController {
@@ -34,32 +36,41 @@ public class CollectionController {
     }
 
     @PostMapping
-    public ResponseEntity<String> create(@RequestBody Collection collection) {
-        collectionService.save(collection);
-        return ResponseEntity.ok().body("La collection a été créée.");
+    public ResponseEntity<String> create(@RequestBody Collection collection, @RequestHeader("Authentication") final String token) {
+        if (verifier(token, Optional.empty())) {
+            collectionService.save(collection);
+            return ResponseEntity.ok().body("La collection a été créée.");
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable("id") final int id) {
-        Optional<Collection> optCollection = collectionService.get(id);
-        if (optCollection.isPresent()) {
-            collectionService.deleteCollection(id);
-            return ResponseEntity.ok().body("La collection a été supprimée.");
+    public ResponseEntity<String> delete(@PathVariable("id") final int id, @RequestHeader("Authentication") final String token) {
+        if (verifier(token, Optional.empty())) {
+            Optional<Collection> optCollection = collectionService.get(id);
+            if (optCollection.isPresent()) {
+                collectionService.deleteCollection(id);
+                return ResponseEntity.ok().body("La collection a été supprimée.");
+            }
+            return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.notFound().build();
+        return ResponseEntity.badRequest().build();
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<String> update(@PathVariable("id") final int id, @RequestBody Collection modified) {
-        Optional<Collection> optCollection = collectionService.get(id);
-        if (optCollection.isPresent()) {
-            Collection current = optCollection.get();
-            if (modified.getName() != null) {
-                current.setName(modified.getName());
+    public ResponseEntity<String> update(@PathVariable("id") final int id, @RequestBody Collection modified, @RequestHeader("Authentication") final String token) {
+        if (verifier(token, Optional.empty())) {
+            Optional<Collection> optCollection = collectionService.get(id);
+            if (optCollection.isPresent()) {
+                Collection current = optCollection.get();
+                if (modified.getName() != null) {
+                    current.setName(modified.getName());
+                }
+                collectionService.save(current);
+                return ResponseEntity.ok().body("La collection " + current.getId() + " a été modifiée.");
             }
-            collectionService.save(current);
-            return ResponseEntity.ok().body("La collection " + current.getId() + " a été modifiée.");
+            return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.notFound().build();
+        return ResponseEntity.badRequest().build();
     }
 }
